fix: resolve ExileCore.Core.ParallelRunner namespace and long-to-int cast on TotalStashes

+++ Core/Adapters/Poe1ShopAdapter.cs
@@ -148,8 +148,8 @@ public class Poe1ShopAdapter : IShopAdapter
             var purchaseWindow = (ingameUi?.PurchaseWindow?.IsVisible == true ? ingameUi.PurchaseWindow : ingameUi?.PurchaseWindowHideout);
             if (purchaseWindow == null) return 1;
 
-            var tabCount = purchaseWindow.TabContainer?.TotalStashes ?? 0;
-            return tabCount > 0 ? tabCount : 1;
+            var tabCount = purchaseWindow.TabContainer?.TotalStashes ?? 0L;
+            return tabCount > 0 ? (int)tabCount : 1;
         }
         catch
         {
@@ -163,7 +163,7 @@ public class Poe1ShopAdapter : IShopAdapter
         {
             var ingameUi = gc?.Game?.IngameState?.IngameUi;
             var purchaseWindow = (ingameUi?.PurchaseWindow?.IsVisible == true ? ingameUi.PurchaseWindow : ingameUi?.PurchaseWindowHideout);
-            return purchaseWindow?.TabContainer?.VisibleStashIndex ?? 0;
+            return (int)(purchaseWindow?.TabContainer?.VisibleStashIndex ?? 0);
         }
         catch
         {
